requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.7
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
idna==2.8
importlib-metadata==0.17
more-itertools==5.0.0
packaging==19.0
paramiko==2.6.0
pluggy==0.12.0
py==1.7.0
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.0
pytest==4.6.2
pytest-html==1.20.0
pytest-metadata==1.8.0
PyYAML==5.1
requests==2.21.0
six==1.12.0
sshtunnel==0.1.5
urllib3==1.24.1
wcwidth==0.1.7
zipp==0.5.1
